Remove spaghetti singleton for CollisionManager

+++ GearShiftLib/GameObjectAdapter.h
@@ -7,8 +7,8 @@ class GameObjectAdapter : public IGameObject {
 public:
 	float getX() const override;
 	float getY() const override;
-	int getWidth() const override;
-	int getHeight() const override;
+	float getWidth() const override;
+	float getHeight() const override;
 	bool isActive() const override;
 	GameObjectAdapter(std::shared_ptr<class GameObject> obj);
